refactor: Extract the Namedpipe Server creation logic to a common class NamedPipeServer.
