+++ src/openjd/adaptor_runtime/_background/background_named_pipe_request_handler.py
@@ -1,29 +1,22 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 
 import json
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, cast
+
+from .._named_pipe import ResourceRequestHandler
 
 if TYPE_CHECKING:  # pragma: no cover because pytest will think we should test for this.
     from .backend_named_pipe_server import WinBackgroundNamedPipeServer
-from openjd.adaptor_runtime._background.named_pipe_helper import (
-    NamedPipeHelper,
-    PipeDisconnectedException,
-)
+
 from openjd.adaptor_runtime._background.server_response import ServerResponseGenerator
-import win32pipe
-import win32file
+
 from pywintypes import HANDLE
-from http import HTTPStatus
 import logging
-import traceback
-
-from openjd.adaptor_runtime._osname import OSName
-
 
 _logger = logging.getLogger(__name__)
 
 
-class WinBackgroundResourceRequestHandler:
+class WinBackgroundResourceRequestHandler(ResourceRequestHandler):
     """
     A handler for managing requests sent to a NamedPipe instance within a Windows environment.
 
@@ -40,65 +33,10 @@ class WinBackgroundResourceRequestHandler:
             server(WinBackgroundNamedPipeServer): The server instance that created this handler.
                 It is responsible for managing the lifecycle of the NamedPipe server and other resources.
                 pipe_handle(pipe_handle): The handle to the NamedPipe instance created and managed by the server.
+            pipe_handle(HANDLE): pipe_handle(HANDLE): Handle for the NamedPipe, established by the instance.
+                Utilized for message read/write operations.
         """
-        if not OSName.is_windows():
-            raise OSError(
-                "WinBackgroundResourceRequestHandler can be only used on Windows Operating Systems. "
-                f"Current Operating System is {OSName._get_os_name()}"
-            )
-        self.server = server
-        self.pipe_handle = pipe_handle
-
-    def instance_thread(self) -> None:
-        """
-        A method that runs in a separate thread to listen to the NamedPipe server. It handles incoming
-        requests and sends back the responses.
-
-        This method calls `send_response` and `handle_request` to process the request and send responses.
-        It should be run in a thread as it continuously listens for incoming requests.
-        """
-        _logger.debug("An instance thread is created to handle communication.")
-        while True:
-            try:
-                request_data = NamedPipeHelper.read_from_pipe(self.pipe_handle)
-                _logger.debug(f"Got following request from client: {request_data}")
-                self.handle_request(request_data)
-            except PipeDisconnectedException as e:
-                # Server is closed
-                _logger.info(f"NamedPipe Server is closed during reading message. {str(e)}")
-                break
-            except Exception:
-                error_message = traceback.format_exc()
-                _logger.error(
-                    f"Encountered an error while reading from the named pipe: {error_message}."
-                )
-                # Try to send back the error message
-                try:
-                    self.send_response(HTTPStatus.INTERNAL_SERVER_ERROR, error_message)
-                except Exception:
-                    _logger.error(
-                        f"Encountered an error while sending the error response: {traceback.format_exc()}."
-                    )
-        try:
-            win32pipe.DisconnectNamedPipe(self.pipe_handle)
-            win32file.CloseHandle(self.pipe_handle)
-        except Exception:
-            _logger.error(
-                f"Encountered an error while closing the named pipe: {traceback.format_exc()}"
-            )
-        _logger.debug("WinBackgroundResourceRequestHandler instance thread exited.")
-
-    def send_response(self, status: HTTPStatus, body: str = ""):
-        """
-        Sends a response back to the client.
-
-        Args:
-            status: An HTTPStatus object representing the HTTP status code of the response.
-            body: A string containing the message body to be sent back to the client.
-        """
-        response = {"status": status, "body": body}
-        _logger.debug(f"NamedPipe Server: Send Response: {response}")
-        NamedPipeHelper.write_to_pipe(self.pipe_handle, json.dumps(response))
+        super().__init__(server, pipe_handle)
 
     def handle_request(self, data: str):
         """
@@ -119,7 +57,10 @@ class WinBackgroundResourceRequestHandler:
             query_string_params = {}
 
         server_operation = ServerResponseGenerator(
-            self.server, self.send_response, body, query_string_params
+            cast("WinBackgroundNamedPipeServer", self.server),
+            self.send_response,
+            body,
+            query_string_params,
         )
         try:
             # TODO: Code refactoring to get rid of the `if...elif..` by using getattr
